refactor all the things
