+++ src/main/java/org/wecancodeit/whendoiboilthewater/models/Recipe.java
@@ -1,4 +1,4 @@
-package org.wecancodeit.whendoiboilthewater.model;
+package org.wecancodeit.whendoiboilthewater.models;
 
 import java.util.ArrayList;
 import java.util.Collection;
