+++ src/main/java/org/wecancodeit/whendoiboilthewater/repositories/IngredientRepository.java
@@ -1,8 +1,10 @@
-package org.wecancodeit.whendoiboilthewater.repository;
+package org.wecancodeit.whendoiboilthewater.repositories;
 
 import org.springframework.data.repository.CrudRepository;
-import org.wecancodeit.whendoiboilthewater.model.Ingredient;
+import org.springframework.stereotype.Repository;
+import org.wecancodeit.whendoiboilthewater.models.Ingredient;
 
+@Repository
 public interface IngredientRepository extends CrudRepository<Ingredient, Long> {
 
 	public Ingredient findByNameAndQuantityIgnoreCase(String name, String quantity);
