+++ src/main/java/org/wecancodeit/whendoiboilthewater/repositories/RecipeRepository.java
@@ -1,0 +1,10 @@
+package org.wecancodeit.whendoiboilthewater.repositories;
+
+import org.springframework.data.repository.CrudRepository;
+import org.springframework.stereotype.Repository;
+import org.wecancodeit.whendoiboilthewater.models.Recipe;
+
+@Repository
+public interface RecipeRepository extends CrudRepository<Recipe, Long>{
+
+}
