+++ src/main/java/org/wecancodeit/whendoiboilthewater/repositories/StepRepository.java
@@ -1,8 +1,10 @@
-package org.wecancodeit.whendoiboilthewater.repository;
+package org.wecancodeit.whendoiboilthewater.repositories;
 
 import org.springframework.data.repository.CrudRepository;
-import org.wecancodeit.whendoiboilthewater.model.Step;
+import org.springframework.stereotype.Repository;
+import org.wecancodeit.whendoiboilthewater.models.Step;
 
+@Repository
 public interface StepRepository extends CrudRepository<Step, Long> {
 
 	Step findBySecBeforeEndAndDescriptionIgnoreCase(Long secondsToEnd, String stepDescription);
