--- src/main/java/org/wecancodeit/whendoiboilthewater/repository/RecipeRepository.java
@@ -1,8 +1,0 @@
-package org.wecancodeit.whendoiboilthewater.repository;
-
-import org.springframework.data.repository.CrudRepository;
-import org.wecancodeit.whendoiboilthewater.model.Recipe;
-
-public interface RecipeRepository extends CrudRepository<Recipe, Long>{
-
-}
