+++ src/main/resources/application.properties
@@ -1,4 +1,4 @@
-spring.datasource.url=jdbc:postgresql://localhost:5432/Pass1234
-spring.jpa.hibernate.ddl-auto=update
-spring.datasource.username=postgres
-spring.datasource.password=Pass1234
+#spring.datasource.url=jdbc:postgresql://localhost:5432/Pass1234
+#spring.jpa.hibernate.ddl-auto=update
+#spring.datasource.username=postgres
+#spring.datasource.password=Pass1234
